backend/requirements.txt
fastapi
uvicorn
python-multipart
Pillow
requests
py7zr
rarfile
pip>=25.3
tokenizers==0.21
transformers==4.55.0
numpy==1.26.4
safetensors
jaconv>=0.3.4
mahotas>=1.4.18
protobuf>=4.25.0
fugashi>=1.3.0
unidic-lite>=1.0.8
sentencepiece>=0.1.99
einops
